+++ NEWS.md
@@ -22,7 +22,9 @@ Following things were added:
   vignettes for this.
 * Reading in of 10x CellRanger h5 files (and multi-file support).
 * NMF for bulk and single cell.
-   
+* [Symphony](https://www.nature.com/articles/s41467-021-25957-x) for label 
+  transfer of cell types.
+* Grouping column for doublet detection methods.
 
 ## Breaking changes
 
@@ -36,7 +38,23 @@ Assertion on 'streaming' failed. Must be of class 'integer', not 'logical'.
 ``` 
 
 Update the streaming parameter to `0L`, `1L` or `2L`, pending on data set size
-and available memory.
+and available memory. Additionally, the interface to the PCA function has
+changed, see here:
+
+```{r}
+calculate_pca_sc(
+  object,
+  no_pcs,
+  pca_params = params_sc_pca(), # new parameter replacing randomised_svd
+  sparse_svd = FALSE,
+  hvg = NULL,
+  seed = 42L,
+  .verbose = TRUE
+)
+```
+
+This gives you way more control over how to run the PCA normalisation (mean
+centering, variance normalisation or now also the new PFLogPF function).
 
 # bixverse 0.3.2